chore: extract filterAndCategorizeUtxos
# Conflicts:
#	suite-common/suite-types/src/device.ts
#	suite-native/analytics/src/events/deviceConnectEvent.ts
#	suite-native/device/src/utils.ts
#	suite-native/firmware/src/components/FirmwareVersionCard.tsx
