+++ suite-common/transaction-search/package.json
@@ -13,8 +13,12 @@
     },
     "dependencies": {
         "@suite-common/metadata-types": "workspace:*",
+        "@suite-common/test-utils": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
-        "@trezor/utils": "workspace:*"
+        "@trezor/blockchain-link": "workspace:*",
+        "@trezor/blockchain-link-types": "workspace:*",
+        "@trezor/utils": "workspace:*",
+        "react": "19.1.0"
     }
 }
