+++ suite-common/transaction-search/src/__fixtures__/filterAndCategorizeUtxosFixtures.ts
