+++ suite-common/transaction-search/src/__tests__/filterAndCategrizeUtxos.test.ts
@@ -2,7 +2,7 @@ import {
     baseUtxo,
     filterAndCategorizeUtxosFixtures,
 } from '../__fixtures__/filterAndCategorizeUtxosFixtures';
-import { filterAndCategorizeUtxos } from '../filterAndCategorizeUtxosUtils';
+import { filterAndCategorizeUtxos } from '../filterAndCategorizeUtxos';
 
 describe(filterAndCategorizeUtxos.name, () => {
     it('filters and categorizes correctly while searching by address', () => {
