+++ suite-common/transaction-search/src/filterAndCategorizeUtxos.ts
