+++ suite-common/transaction-search/src/index.ts
@@ -1,1 +1,5 @@
 export { advancedSearchTransactions } from './advancedSearchTransactions';
+export { filterAndCategorizeUtxos } from './filterAndCategorizeUtxos';
+export { useFilteredUtxos } from './useFilteredUtxos';
+export { useExcludedUtxos } from './useExcludedUtxos';
+export { getExcludedUtxos } from './getExcludedUtxos';
