+++ suite-common/transaction-search/src/useExcludedUtxos.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 
 import { Account } from '@suite-common/wallet-types';
 
-import { GetExcludedUtxosProps, getExcludedUtxos } from '../sendFormUtils';
+import { GetExcludedUtxosProps, getExcludedUtxos } from './getExcludedUtxos';
 
 interface UseExcludedUtxosProps extends GetExcludedUtxosProps {
     account: Account;
