+++ suite-common/transaction-search/tsconfig.json
@@ -4,8 +4,15 @@
     "include": [".", "**/*.json"],
     "references": [
         { "path": "../metadata-types" },
+        { "path": "../test-utils" },
         { "path": "../wallet-types" },
         { "path": "../wallet-utils" },
+        {
+            "path": "../../packages/blockchain-link"
+        },
+        {
+            "path": "../../packages/blockchain-link-types"
+        },
         { "path": "../../packages/utils" }
     ]
 }
