+++ suite-common/wallet-utils/src/index.ts
@@ -27,9 +27,6 @@ export * from './transactionUtils';
 export * from './validationUtils';
 export * from './AmountTypes';
 export * from './baseCurrency';
-export * from './filterAndCategorizeUtxosUtils';
-export * from './hooks/useExcludedUtxos';
-export * from './hooks/useFilteredUtxos';
 export * from './cardanoStakingUtils';
 export * from './amountUtils';
 export * from './bigNumberUtils';
