+++ suite-common/wallet-utils/tsconfig.json
@@ -4,7 +4,6 @@
     "include": [".", "**/*.json"],
     "references": [
         { "path": "../fiat-services" },
-        { "path": "../metadata-types" },
         { "path": "../suite-constants" },
         { "path": "../suite-types" },
         { "path": "../suite-utils" },
@@ -15,9 +14,6 @@
         { "path": "../../suite/intl" },
         {
             "path": "../../packages/address-validator"
-        },
-        {
-            "path": "../../packages/blockchain-link"
         },
         {
             "path": "../../packages/blockchain-link-types"
